Don't publish this package under different names because of the following:
@enkore-target/js-none, @enkore-target/js-node etc. all exist because:

- we need to have additional packages installed for js-node and js-web (@types/node and @types/web).
This would be messy to implement with a single package.

- the end-user can interact with this package, i.e. it's a public API.

The enkore JS toolchain is not strictly a public API, it will be made available to the end-user
for convenience but no guarantees about the API will be made - this is why this API is not described
in @enkore/spec.

There's simply no reason to separate the toolchain into multiple packages as of right now.

Using a single toolchain package for JS projects only has one real downside:

The target API needs to tell enkore what toolchain it wants to use.

While the upsides are:

- only one types package, which makes reading the source code of @enkore/target-js-factory **much** easier to understand:

target-js-factory is the blueprint for js-node, js-none and js-web targets, but typing information is static, which
toolchain types package would the factory use (if we had different toolchain packages)? All of them would be suitable,
since they are copies of eachother, but it surely would be weird to read. (e.g. using js-node toolchain type for a package
that is the blueprint for js-none, js-web and js-node).

- IF we decide to make the toolchain API public, it would be easier for the end-user to use, they would only
need to install one toolchain types package.